+++ internal/gopher-lua/parse/lexer.go
@@ -4,7 +4,7 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"github.com/yuin/gopher-lua/ast"
+	"github.com/heroiclabs/nakama/v2/internal/gopher-lua/ast"
 	"io"
 	"reflect"
 	"strconv"